+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Custom hot key
 - Force boot scene on play
 - Create domain pattern
+- Create common module template